Import every root pages from workspace
